Added: GPS Location and Network Location             (Not working on emulator)             Yelp search by Restaurant by ID method Change Icon label Integrate methods to ResultFragment Removed Result Activity (display result in viewpager) Swapped Restaurant argument for restaurant and rating image Modified MainActivity

+++ app/src/main/java/app/team3/t3/ResDatabaseHelper.java
@@ -262,8 +262,8 @@ public class ResDatabaseHelper extends SQLiteOpenHelper {
                 cursor.getString(7),
                 cursor.getString(8),
                 Integer.parseInt(cursor.getString(9)),
-                Float.parseFloat(cursor.getString(10)),
-                Float.parseFloat(cursor.getString(11)),
+                Double.parseDouble(cursor.getString(10)),
+                Double.parseDouble(cursor.getString(11)),
                 cursor.getString(12),
                 cursor.getString(13));
 
@@ -294,8 +294,8 @@ public class ResDatabaseHelper extends SQLiteOpenHelper {
                 cursor.getString(7),
                 cursor.getString(8),
                 Integer.parseInt(cursor.getString(9)),
-                Float.parseFloat(cursor.getString(10)),
-                Float.parseFloat(cursor.getString(11)),
+                Double.parseDouble(cursor.getString(10)),
+                Double.parseDouble(cursor.getString(11)),
                 cursor.getString(12),
                 cursor.getString(13));
